Fix Matter WindowCovering control end-to-end
Four issues prevented HA Matter commands from actually moving the vent:

1. WindowCovering cluster requires a Delegate to handle UpOrOpen,
   DownOrClose, GoToLiftPercentage, and StopMotion. Without it the
   cluster logs "no delegate set" and ignores the command. Added
   VentCoveringDelegate that reads the Target attribute and calls the
   existing position callback into Rust. Register it AFTER
   esp_matter::start() so the cluster server instance exists.

2. The cluster only invokes HandleMovement when Feature::kPositionAwareLift
   is set in the FeatureMap. esp_matter's window_covering_device default
   is ESP_MATTER_NONE_FEATURE_ID. Explicitly add Lift +
   PositionAwareLift features after creating the endpoint.

3. percent100ths_to_angle and angle_to_percent100ths overflowed u16 when
   multiplying by 10000 or by the range (90). e.g., pct=10000 yielded
   angle=176 instead of 90, so close stopped at 176deg. Promote the
   intermediate arithmetic to u32.

4. CHIP/Matter init starvation on the C6 was hitting the 5s task
   watchdog (long-dispatch warnings up to 3.7s). Bumped
   CONFIG_ESP_TASK_WDT_TIMEOUT_S to 30 and disabled
   CONFIG_CHIP_DEVICE_CONFIG_ENABLE_WIFI_STATION (this is a Thread-only
   Matter device) to reduce startup work.

+++ firmware/vent-controller/src/matter.rs
@@ -33,16 +33,16 @@ extern "C" {
 /// In Matter, 0% = open, 100% = closed.
 pub fn angle_to_percent100ths(angle: u8) -> u16 {
     let clamped = angle.clamp(ANGLE_CLOSED, ANGLE_OPEN);
-    let range = (ANGLE_OPEN - ANGLE_CLOSED) as u16; // 90
-    let from_open = (ANGLE_OPEN - clamped) as u16;
-    (from_open * 10000) / range
+    let range = (ANGLE_OPEN - ANGLE_CLOSED) as u32; // 90
+    let from_open = (ANGLE_OPEN - clamped) as u32;
+    ((from_open * 10000) / range) as u16
 }
 
 /// Convert Matter percent100ths (0–10000) to servo angle (90–180).
 /// In Matter, 0% = open, 100% = closed.
 pub fn percent100ths_to_angle(pct: u16) -> u8 {
-    let clamped = pct.min(10000);
-    let range = (ANGLE_OPEN - ANGLE_CLOSED) as u16; // 90
+    let clamped = pct.min(10000) as u32;
+    let range = (ANGLE_OPEN - ANGLE_CLOSED) as u32; // 90
     let from_open = (clamped * range) / 10000;
     ANGLE_OPEN - from_open as u8
 }
